+++ playground/backend/database/migrations/1742032741448_create_note_tags_table.ts
@@ -1,8 +1,8 @@
-import { castle } from '#config/castle'
+import { dbRef } from '#config/database'
 import { BaseSchema } from '@adonisjs/lucid/schema'
 
 export default class extends BaseSchema {
-  protected tableName = castle.table.noteTags()
+  protected tableName = dbRef.table.noteTags()
 
   async up() {
     this.schema.createTable(this.tableName, (t) => {
@@ -10,13 +10,13 @@ export default class extends BaseSchema {
 
       t.integer('note_id')
         .unsigned()
-        .references(castle.table.note('id'))
+        .references(dbRef.table.note('id'))
         .notNullable()
         .onDelete('CASCADE')
 
       t.integer('tag_id')
         .unsigned()
-        .references(castle.table.tag('id'))
+        .references(dbRef.table.tag('id'))
         .notNullable()
         .onDelete('CASCADE')